moved code to api util function
